Enhance routing and layout structure for protected components
- Introduced a new `Layout` component to wrap protected routes, providing a consistent layout across the application.
- Updated `RouteWrapper` in `App.js` to include the `Layout` component for all protected routes, ensuring a unified user experience.
- Modified `Home` component to utilize the new `Layout` for the `AuthenticatedDashboard`.
- Cleaned up the `Profile` and `AuthenticatedDashboard` components by removing redundant header elements, improving overall code clarity and maintainability.
